+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # lgLib utility plugin for glFusion
 
+## 1.1.0
+Release TBD
+  * Add admin interface for job queue.
+  * Add started timestamp field to job queue table.
+
 ## 1.0.13
 Release 2021-07-17
   * Only use session ID in message table for anonymous recipients.